moves io modules to csv modules outside of database_modules. Creates new routers directory for API endpoints and main.py for app entry point. Removes old config file.
