+++ backend/tests/test_database_modules/test_io/test_transactions_csv_loader.py
@@ -9,7 +9,7 @@ from datetime import datetime
 import os
 
 # Local imports
-from backend.database_modules.io.transactions_csv_loader import generate_base_hash, iter_val_csv_file
+from backend.csv_modules.transactions_csv_loader import generate_base_hash, iter_val_csv_file
 from backend.database_modules.models.transactions import transaction_columns
 from backend.tests.conftest import TEST_CSV_DIR
